DEV-172 update for elasticsearch 7
Update mappings to work with Elasticsearch 7. Some of these changes
are backwards-incompatible, so this also updates the mappings to _stop_
working with ES5.

Remove the explicit setting to disable the `_all` field. ES7 no
longer supports `_all` at all, so we can't even say to disable it.

Set `index.mapping.nested_objects.limit` to a ridiculously huge value.
This is a new setting in ES7 that would otherwise prevent us from
storing the tens of thousands of nested documents we sometimes need.

Make setuptools_scm stop complaining if something else tries to pip
install this library with `-e`.