DEV-327 update gene expression settings
The query to the gene expression index is optimized for performance
and uses inner hits to query subdocuments for genes. Since there's
20k genes, I'm setting this config to 20k.  It defaults to 100,
which is too small for our expected usage patterns. We also don't
expect 20k as regular usage patterns.  But we'll have throttles
in gdcapi to limit the total number of documents returned. So,
this should not cause an operational issue.